fix: add disabled handling to real acc opening too
